Increase tapping term since the lower value was causing missed taps
Sometimes I would hold the a key when doing ctrl+a for example and I
would get the layer switch on the a key when I wanted the a. Since the
alphas are more important than the layer switch I increased the number
to favor alphas which will increase the errors when layer switching
which is done less often so I'm ok with this trade off.